+++ src/main/java/com/geocode/octowallet/entities/CasualIncome.java
@@ -1,5 +1,6 @@
 package com.geocode.octowallet.entities;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import javax.persistence.Entity;
@@ -15,7 +16,7 @@ public class CasualIncome extends Income {
   public CasualIncome() {
   }
 
-  public CasualIncome(Long id, String description, Double value, LocalDate date) {
+  public CasualIncome(Long id, String description, BigDecimal value, LocalDate date) {
     super(id, description, value, date);
   }
